+++ src/main/java/com/schwab/urlshortener/exception/ShortCodeGenerationException.java
@@ -1,0 +1,8 @@
+package com.schwab.urlshortener.exception;
+
+public class ShortCodeGenerationException extends RuntimeException {
+
+    public ShortCodeGenerationException(String message) {
+        super(message);
+    }
+}