+++ src/main/java/com/schwab/urlshortener/util/ShortCodeGenerator.java
@@ -1,0 +1,6 @@
+package com.schwab.urlshortener.util;
+
+public interface ShortCodeGenerator {
+
+    String generate(int length);
+}